+++ app/controllers/stellar_base/withdraw_controller.rb
@@ -1,16 +1,8 @@
 module StellarBase
   class WithdrawController < ApplicationController
 
-    WITHDRAWAL_REQUEST_PARAMS = %i[
-      type
-      asset_code
-      dest
-      dest_extra
-      fee_network
-    ]
-
     def create
-      op = WithdrawalRequests::Operations::Create.(withdrawal_request: withdrawal_request_params)
+      op = WithdrawalRequests::Operations::Create.(withdrawal_request: params)
 
       respond_to do |f|
         f.json do
@@ -19,16 +11,10 @@ module StellarBase
             representer = WithdrawalRequestRepresenter.new(twin)
             render json: representer
           else
-            head :unprocessable_entity
+            render json: {error: op["contract.default"].errors}, status: :unprocessable_entity
           end
         end
       end
-    end
-
-    private
-
-    def withdrawal_request_params
-      params.permit(*WITHDRAWAL_REQUEST_PARAMS)
     end
 
   end